Added intitial models/views/url config.
Needs work/testing.  4 TODOs in views should be highest priority
Testing should occur before working further on TODOs
Static files need to be setup before testing can occur
Recommended testing on a different db on this branch.  Pushing to
'Backend' can be saved for when it is confirmed working and ready
to use MySQL db.