Added new module for auth. Started working on POST
started working on post.  Still needs a bunch of
testing.  Also in order to test, you will need a
csrf token.  I have been testing with curl, and
have made an easy way to get a csrf token with
the puck_user app.  Just hit
localhost:8000/authenticate and it should give
you one.  This will be the start of our
authentication