Merged from Hopper. Verification needed for db
This needs to be put on AWS server and checked to make sure I didn't
break the MySQL database.  It should be good (I think).  After
verification, it can be merged up to master, and then all the django
code for initial GET/POST methods should be complete, and ready for
integration, and future iterations.